Add support for server-side language bindings
This commit adds support for server-side languages bindings.  This is done in
two ways:

1. Adding versions of wl_client_add_object and wl_client_new_object that take
a dispatcher and dispatcher-specific implementation pointer.  This allows for
runtime calling of native language functions for callbacks instead of having to
generate function pointers.

2. Adding versions of wl_resource_post_event and wl_resource_queue_event that
take an array of wl_argument instead of a variable argument list.  This allows
for easier run-time argument conversion and removes the need for libffi-based
calling of variadic functions.